+++ e2e/test_bash.py
@@ -120,6 +120,17 @@ def test_specializes_for_bash(complgen_binary_path: Path):
         assert get_sorted_bash_completions(path, input) == sorted(["bash"])
 
 
+def test_specializes_for_bash_with_regex(complgen_binary_path: Path):
+    GRAMMAR = (
+        """cmd <FOO>bar; <FOO> ::= {{{ echo foo }}}; <FOO@bash> ::= {{{ echo bash }}}@bash"bash";"""
+    )
+    with completion_script_path(complgen_binary_path, GRAMMAR) as path:
+        input = (
+            r'''COMP_WORDS=(cmd bash); COMP_CWORD=1; _cmd; printf '%s\n' "${COMPREPLY[@]}"'''
+        )
+        assert get_sorted_bash_completions(path, input) == sorted(["bashbar"])
+
+
 def test_nontail_matching_alternative(complgen_binary_path: Path):
     GRAMMAR = """cmd <LEFT> | right; <LEFT> ::= {{{ echo left }}}@bash"left";"""
     with completion_script_path(complgen_binary_path, GRAMMAR) as path: