+++ src/main.rs
@@ -247,7 +247,7 @@ fn aot(args: &Cli) -> anyhow::Result<()> {
             .context(regex_dot_file_path.clone())?;
     }
 
-    let dfa = match DFA::from_regex(shell, regex, validated.subdfas) {
+    let dfa = match DFA::from_regex(regex, validated.subdfas) {
         Ok(dfa) => dfa,
         Err(e) => return handle_error(e, &input, &validated.command),
     };