+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+ - Specialization command now also support regexes
  - Added `--regex` option for writing generated regular expression into a .dot file for debugging
  - More precise ambiguity detector: Detect clashing subword leaders